non-optimal fix for GROOVY-1809: non-static fields should not be treated as static properties
git-svn-id: http://svn.codehaus.org/groovy/trunk/groovy/groovy-core@8525 a5544e8c-8a19-0410-ba12-f9af4593a198

+++ src/test/groovy/ArrayTest.groovy
@@ -128,7 +128,7 @@ class ArrayTest extends GroovyTestCase {
         Object[] y = [[1,2]]
         assert y == x
         assert [[1,2],[3,4]] as int[][] == [[1,2],[3,4]] as int[][]
-        assert [[[5,6],[7,8]]] as int[][][] == [[[5,6],[7,8]]] as int[][][]
+        assert [[[5,6],[7,8]]] as int[][][] == [[[5,6],[7,8]]] as Long[][][]
         assert [[1,2],[3,4]] as long[][] == [[1,2],[3,4]] as long[][]
         assert [[1,2],[3,4]] as long[][] == [[1,2],[3,4]] as Long[][]
         assert [[1,2],[3,4]] as long[][] == [[1,2],[3,4]]
